Create folder structure and empty README.md
In master-react, create a good folder structure in which all
learnings are categorized by topics and the material referred.

Create a sub-folder "react.dev-docs" in master-react directory which
will contain everything from notes to react apps, exercises and
examples that will be studied using the React.dev Docs.

In sub-folder "react.dev-docs", create folders for topics and notes.
    - Topics folder will contain all the topics (4 broad categories in
      the official docs). These should be created using
      create-react-app.
    - Notes folder will similarly contain .txt files for all the
      create-react-app sub-folders (that will be created as the course
      progresses) and all notes for their respective topics should be
      written in their respective .txt files having the same name.