Study conditional rendering by making a to-do-list
In subtopics, create a folder 6-conditional-rendering
and study this topic by making a to-do list.

Study different syntax uses, such as if..else, ternary
operators and the approach of creating a new variable and
then using if..else to conditionally render components.

Update the import in App.js so that the List component
is rendered.