Study "Rendering-lists" by creating Gallery of scientists
In subtopics, create a new folder 7-rendering-lists to study
the topic of Rendering Lists in React. Create 1-scientists-profile.js
to create a component Gallery which renders a gallery of scientists.

Save a list of scientists in scientists.js to import in 1-scientists-
profile.js to use this list to render each scientist's profile.

Import this new Gallery component in App.js