Build functionality to add items to the cart
Create cart-item component and write the logic that adds the
functionality of adding products to cart. The cart should only add
the product to it, if the added product doesn't exist as a cart-item
and if the product already exists, then it should update its quantity.

Build this using context in CartContext