Create the checkout component and its core functionality
Create a checkout component that will take users to the checkout page
once they click the checkout button on the checkout dropdown.

Build the main functionality that allows user to see all the products
added to cart at one place and allows them to change the quantity of
a particular product from that page only.

The checkout page should also reflect the total price of all the items
in the cart.