Refactor products.context.jsx
Replace "product" with "categories" in product.context.jsx. Rename the
file to categories.context.jsx because categories is more meaningful
and relevant to the app and the custom metods creted inside
firebase.utils.js.

Refactor shop component to utilize this new category map and display
all categories and their items on the shop page.